Divide actions to separated files
