+++ webapp/package.json
@@ -44,6 +44,7 @@
     "@babel/core": "^7.4.5",
     "@babel/plugin-proposal-class-properties": "7.4.4",
     "@babel/plugin-syntax-dynamic-import": "7.2.0",
+    "@babel/polyfill": "^7.6.0",
     "@babel/preset-env": "^7.4.5",
     "@babel/preset-react": "^7.0.0",
     "@babel/runtime": "^7.4.5",
