+++ webapp/src/components/post_type/action_view/index.js
@@ -3,7 +3,7 @@ import {bindActionCreators} from 'redux';
 
 import {getConfig} from 'mattermost-redux/selectors/entities/general';
 
-import {fetchPollMetadata} from 'actions/vote';
+import {fetchPollMetadata} from 'actions/poll_metadata';
 import {pollMetadata} from 'selector';
 
 import ActionView from './action_view';
