+++ webapp/src/plugin.jsx
@@ -1,5 +1,5 @@
 import {configurationChange, fetchPluginConfiguration} from 'actions/config';
-import {websocketHasVoted} from 'actions/vote';
+import {websocketHasVoted} from 'actions/poll_metadata';
 
 import {id as pluginId} from './manifest';
 import reducer from './reducers';
