+++ webapp/src/reducers/index.js
@@ -1,7 +1,7 @@
 import {combineReducers} from 'redux';
 
 import {postTypeComponent} from './post_type';
-import {pollMetadata} from './vote';
+import {pollMetadata} from './poll_metadata';
 
 export default combineReducers({
     postTypeComponent,
