+++ webapp/tests/setup.js
@@ -1,3 +1,4 @@
+import '@babel/polyfill';
 import Adapter from 'enzyme-adapter-react-16';
 import {configure} from 'enzyme';
